+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Close button variant. 
+
 ## 0.1.1 - 2025-12-17
+
+### Added
 
 - Initial release.